relative error check with doubles

+++ test/common/correctness_helpers.h
@@ -3,8 +3,8 @@
 #include <algorithm>
 #include <math.h>
 
-bool check_relative_error_okay(float result, float expected, float epsilon) {
-    auto error = fabsf(result - expected);
-    auto rel_error = error / std::max(1.f, fabsf(expected));
+bool check_relative_error_okay(double result, double expected, double epsilon) {
+    auto error = std::fabs(result - expected);
+    auto rel_error = error / std::max((double)1.0, std::fabs(expected));
     return rel_error < epsilon;
 }
